Revert back to where things worked
This is a combination of 3 commits.
The first commit's message is:

Revert "Merge pull request #5 from eHealthAfrica/tim/docker-travis"

This reverts commit 50ea0593b69b7ad741917b5a001acbb50ebc39f2, reversing
changes made to 54cad68ae705a413b53e1157d0df8e78e77d98ea.

Revert "Reorganized folder structure and fix docker"

This reverts commit 54cad68ae705a413b53e1157d0df8e78e77d98ea, reversing
changes made to 88ba6780cd6c97c0914a7b0e426d77285a0d0b23.

Revert "Reorganized folder structure and fix docker"

This reverts commit 88ba6780cd6c97c0914a7b0e426d77285a0d0b23.
