+++ README.md
@@ -5,18 +5,12 @@
 [![Build Status](https://travis-ci.org/eHealthAfrica/gather2.svg?branch=master)](https://travis-ci.org/eHealthAfrica/gather2)
 
 
-## Requirements
-
-```
-sudo apt-get install python3-dev
-```
-
 ## setup
 
 ```
-pyvenv-3.4 ./gather_env
 git clone git@github.com:eHealthAfrica/gather2.git
 cd gather2
+virtualenv-3.4 .
 bin/pip install -r dev-requirements.txt
 bin/invoke runserver
 ```
