--- setup.cfg
@@ -1,4 +1,0 @@
-[flake8]
-ignore = E501,C901
-max-line-length = 100
-exclude = migrations
