+++ src/gather2_core/XForm2JSONSchema.py
@@ -1,0 +1,11 @@
+def walk(d):
+    ret = {}
+    if 'children' in d:
+        ret = {}
+        ret["properties"] = dict((c['name'], walk(c)) for c in d['children'])
+    _type = d.get('type', None)
+    if _type == 'string':
+        ret = {}
+        ret['type'] = 'string'
+        ret['description'] = d['label']
+    return ret
