+++ src/gather2_core/manage.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python
+import os
+import sys
+
+if __name__ == "__main__":
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "gather2_core.settings")
+
+    from configurations.management import execute_from_command_line
+
+    execute_from_command_line(sys.argv)
