+++ src/gather2_core/templates/surveyviewset_list.html
