+++ src/gather2_core/tests.py
@@ -1,0 +1,9 @@
+import pytest
+
+from .models import Survey
+
+
+@pytest.mark.django_db
+def test_my_user():
+    a = Survey(name="Hello World")
+    assert str(a) == "None - Hello World"
